feat: Ticket정보 조회 api UserApi로 변경
