+++ application/ticket-app-api/src/main/java/com/uket/app/ticket/api/controller/UserApi.java
@@ -4,9 +4,11 @@ import com.uket.app.ticket.api.dto.request.UserRegisterRequest;
 import com.uket.app.ticket.api.dto.response.AuthResponse;
 import com.uket.core.dto.response.ErrorResponse;
 import com.uket.domain.auth.config.userid.LoginUserId;
+import com.uket.domain.ticket.dto.CheckTicketDto;
 import com.uket.domain.user.dto.UserInfoDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -14,8 +16,10 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
+import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -76,5 +80,19 @@ public interface UserApi {
     ResponseEntity<UserInfoDto> getUserInfo(
             @Parameter(hidden = true)
             @LoginUserId Long userId
+    );
+
+    @Operation(summary = "사용자 티켓 조회 API", description = "특정 사용자의 모든 티켓을 조회합니다.")
+    @GetMapping("/tickes")
+    @ApiResponse(responseCode = "200", description = "OK", content = @Content(array = @ArraySchema(schema = @Schema(implementation = CheckTicketDto.class))))
+    @ApiResponse(responseCode = "404", description = "NOT FOUND", content = @Content(mediaType = "application/json",
+        examples = @ExampleObject(name = "US0001", description = "토큰에 담긴 UserId에 대한 사용자를 찾을 수 없을 때 발생합니다.",
+            value = """
+                        {"code": "US0001", "message": "해당 사용자를 찾을 수 없습니다."}
+                        """
+        ), schema = @Schema(implementation = ErrorResponse.class)))
+    List<CheckTicketDto> getUserTickets(
+        @Parameter(hidden = true)
+        @LoginUserId Long userId
     );
 }
