+++ application/ticket-app-api/src/main/java/com/uket/app/ticket/api/controller/impl/UserController.java
@@ -3,12 +3,16 @@ package com.uket.app.ticket.api.controller.impl;
 import com.uket.app.ticket.api.controller.UserApi;
 import com.uket.app.ticket.api.dto.request.UserRegisterRequest;
 import com.uket.app.ticket.api.dto.response.AuthResponse;
+import com.uket.app.ticket.api.service.TicketingService;
 import com.uket.app.ticket.api.service.UserRegisterService;
+import com.uket.domain.auth.config.userid.LoginUserId;
 import com.uket.domain.auth.dto.response.AuthToken;
+import com.uket.domain.ticket.dto.CheckTicketDto;
 import com.uket.domain.user.dto.CreateUserDetailsDto;
 import com.uket.domain.user.dto.UserInfoDto;
 import com.uket.domain.user.entity.Users;
 import com.uket.domain.user.service.UserService;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -21,6 +25,7 @@ public class UserController implements UserApi {
 
     private final UserService userService;
     private final UserRegisterService userRegisterService;
+    private final TicketingService ticketingService;
 
     @Override
     public ResponseEntity<AuthResponse> register(Long userId, UserRegisterRequest request) {
@@ -39,6 +44,11 @@ public class UserController implements UserApi {
          UserInfoDto userInfoDto = userService.getUserInfo(userId);
 
         return ResponseEntity.ok(userInfoDto);
+    }
+
+    @Override
+    public List<CheckTicketDto> getUserTickets(Long userId) {
+        return ticketingService.checkUserTickets(userId);
     }
 
     private CreateUserDetailsDto generateCreateUserDetailsDto(UserRegisterRequest request) {
